# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prime Max & Elite Prefab ERP CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # DEPARTMENTS
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  # USERS
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # DEALS
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    put:
      operationId: updateDeal
      tags: [deals]
      summary: Update deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "200":
          description: Deal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ACTIVITIES
...
  /departments/{id}:
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        prefix:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        trn:
          type: string
        vatPercent:
          type: number
        logo:
          type: string
        bankDetails:
          type: string
        letterhead:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, shortName, prefix, isActive, createdAt]
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        prefix:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        trn:
          type: string
        vatPercent:
          type: number
        logo:
          type: string
        bankDetails:
          type: string
        letterhead:
          type: string
      required: [name, shortName, prefix]
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, isActive, createdAt]
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
...
    UpdateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        companyName:
          type: string
        designation:
          type: string
        companyId:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        companyName:
          type: string
        designation:
          type: string
        companyId:
          type: integer
        notes:
          type: string
      required: [name]
    Deal:
      type: object
      properties:
        id:
          type: integer
        dealNumber:
          type: string
        title:
          type: string
        clientName:
          type: string
        value:
          type: number
        stage:
          type: string
        probability:
          type: number
        expectedCloseDate:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        companyId:
          type: integer
        companyRef:
          type: string
        leadId:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, dealNumber, title, stage, createdAt]
    CreateDealBody:
      type: object
      properties:
        title:
          type: string
        clientName:
          type: string
        value:
          type: number
        stage:
          type: string
        probability:
          type: number
        expectedCloseDate:
          type: string
        assignedToId:
          type: integer
        companyId:
          type: integer
        leadId:
          type: integer
        notes:
          type: string
      required: [title, stage]
...